+++ controllers/email.js
@@ -19,6 +19,7 @@ router.post('/', async (req, res) => {
 		const success = await sgMail.send(msg);
 		res.status(200).end();
 	} catch (error) {
+		console.log(error);
 		res.status(500).json({
 			message: 'Sorry but something went wrong while sending the emails.'
 		});
