+++ helpers/slackConfirmation.js
